Added styling to event list page.

+++ FierceLadies/events/views.py
@@ -5,7 +5,7 @@ from django.views.generic.edit import CreateView, UpdateView, DeleteView
 from django.db.models import Q
 from django.urls import reverse_lazy
 from .forms import EventForm
-
+from django.contrib.auth.decorators import login_required
 
 from .models import Event, Comment
 # Create your views here.
@@ -27,10 +27,12 @@ def eventPage(request, pk):
     }
     return render(request, 'events/event_detail.html', context)
 
-
+@login_required(login_url='login')
 def deleteComment(request, pk):
     comment = Comment.objects.get(id=pk)
     event = comment.event
+    if comment.user != request.user:
+        return render(request, 'unauthorized_access.html')
     if request.method == 'POST':
         comment.delete()
         return redirect('event-detail', event.id)
@@ -41,7 +43,6 @@ def deleteComment(request, pk):
         }
         return render(request, 'events/comment_delete_confirmation.html', context)
 
-
 def eventList(request):
     search_area = request.GET.get('search-area')
     if search_area is None:
@@ -49,11 +50,12 @@ def eventList(request):
     events = Event.objects.filter(Q(name__icontains=search_area) | Q(
         description__icontains=search_area) | Q(location__icontains=search_area))
     context = {
-        'events': events
+        'events': events,
+        'search_area' : search_area,
     }
     return render(request, 'events/event_list.html', context)
 
-
+@login_required(login_url='login')
 def createEvent(request):
     if request.method == 'POST':
         form = EventForm(request.POST)
@@ -68,9 +70,11 @@ def createEvent(request):
         }
         return render(request, 'events/event_form.html', context)
 
-
+@login_required(login_url='login')
 def updateEvent(request, pk):
     event = Event.objects.get(id=pk)
+    if event.host != request.user:
+        return render(request, 'unauthorized_access.html')
     if event.host != request.user:
         return redirect('event-detail', event.id)
     elif request.method == 'POST':
@@ -85,7 +89,6 @@ def updateEvent(request, pk):
             'form': form,
         }
         return render(request, 'events/event_form.html', context)
-
